Create glutton.cpp
this just sorts the entered numbers. next time, make ar an array of structs instead of ints to keep track of who ate what, but then we'll have to change quicksort too. 
Also, still have to ask the user how many people are eating pancakes so it's not just 10. 
If want to, finally, can add checks to make sure user enters valid numbers